backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.6
pydantic-settings==2.7.1
python-multipart==0.0.20
orjson==3.10.15

# ingestion
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
pdfminer.six==20231228
docx2txt==0.8

# ML
numpy==2.1.3
scikit-learn==1.5.2
joblib==1.4.2

# tests
pytest==8.3.4

python-docx==1.1.2
